requirements.txt
pixivpy3
discord.py